REF: drop over-engineered tests/extraction; keep minimal tap fix
Jest cannot prove transparent-header overlay hit-testing. Revert the
component extraction and prop-parroting unit suite. Keep only the
production pointerEvents passthrough and larger headerRight hit target.
